// frontend/src/ipc/generated/RelatedTaskView.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { RelatedTaskKind } from "./RelatedTaskKind";

export type RelatedTaskView = { id: string, kind: RelatedTaskKind, title: string, group: string, local_date: string | null, series_id: string | null, };